refactor: organize profile components into their respective directories
