+++ frontend/src/components/profile/FavouritesPreview.tsx
@@ -1,4 +1,4 @@
-// src/components/FavouritesPreview.tsx
+// src/components/profile/FavouritesPreview.tsx
 
 'use client'
